Migrate backend models and API endpoints to new dragon schema

+++ src/main/java/org/itmo/isLab1/coordinates/dto/CoordinateDto.java
@@ -8,6 +8,6 @@ import org.itmo.isLab1.common.framework.dto.CrudDto;
 @EqualsAndHashCode(callSuper = true)
 public class CoordinateDto extends CrudDto {
   private int id;
-  private Integer x;
-  private Double y;
+  private Float x;
+  private Long y;
 }
